Create Computer object that can play code maker or breaker
Gave Computer objects an extra instance variable called
@options. This is four nested arrays of code options that
are randomly selected by the computer to create its own code.

Added a method that eliminates past guesses from @options as
the game progresses and keeps exact matches.

As the game progresses, the computer will 'get smarter' as it
narrows down on the correct code.